# docs/swagger.yaml
basePath: /api
definitions:
  handlers.UpdateUserRolesRequest:
    properties:
      role_ids:
        items:
          type: integer
        type: array
    required:
    - role_ids
    type: object
...
  response.Response:
    properties:
      code:
        description: Business status code
        type: integer
      data:
        description: Response data
      message:
        description: Response message
        type: string
      trace_id:
        description: Trace ID for request tracking
        type: string
    type: object
...
  ws.Message:
    properties:
      content:
        type: string
      from:
        type: string
      timestamp:
        type: integer
      to:
        description: User ID or Group ID
        type: string
      type:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: A modern Go admin scaffold API server.
  title: Go Admin Scaffold API
  version: "1.0"
paths:
...
  /admin/v1/users/{id}/roles:
    put:
      consumes:
      - application/json
      description: Updates the roles assigned to a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Role IDs
        in: body
        name: roles
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateUserRolesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Update user roles
      tags:
      - User
  /ws:
    get:
      consumes:
      - application/json
      description: Establishes a WebSocket connection for real-time chat
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      - description: JWT Token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "101":
          description: Switching Protocols to websocket
          schema:
            type: string
      summary: Connect to WebSocket
      tags:
      - WebSocket
  /ws/join:
    post:
      consumes:
      - application/json
      description: Adds a user to a chat group
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      - description: Group ID
        in: query
        name: group_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Join Chat Group
      tags:
      - WebSocket
  /ws/leave:
    post:
      consumes:
      - application/json
      description: Removes a user from a chat group
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      - description: Group ID
        in: query
        name: group_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Leave Chat Group
      tags:
      - WebSocket
  /ws/send:
    post:
      consumes:
      - application/json
      description: Sends a message (private, group, or announcement)
      parameters:
      - description: Message Object
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/ws.Message'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Send Message
      tags:
      - WebSocket
swagger: "2.0"